Fix build issues occurred when engine/lotus points to ORT Github latest

+++ onnxruntime/core/providers/dml/DmlExecutionProvider/src/Operators/OperatorUtility.cpp
@@ -219,7 +219,7 @@ namespace Dml
                 return std::nullopt;
             }
 
-            if (opIt->inputCountFilter && *opIt->inputCountFilter != candidateOpInputCount)
+            if (opIt->inputCountFilter && *opIt->inputCountFilter != static_cast<uint32_t>(candidateOpInputCount))
             {
                 return std::nullopt;
             }
